requirements.txt
matplotlib==3.9.3
numpy==1.26.3
opencv-python==4.10.0.84
pillow==10.2.0
propcache==0.2.1
pycocotools==2.0.8
pysodmetrics==1.4.2
pytorch_ssim==0.1
seaborn==0.13.2
shapely==2.0.6
sympy==1.13.1
tensorboard==2.19.0
timm==1.0.14
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.1.2+cu121
torchmetrics==1.6.0
torchvision==0.16.2+cu121
tqdm==4.67.1
h5py==2.10.0

